requirements.txt
Cython
numpy
Pillow
torch
torchvision
matplotlib
pycocotools
opencv-python
imutils
requests